character animation/movement changes
changed the way the character moves and displays the animations. There are now two booleans that can restrict movement. allowCharMove controls weather the character can move at all. If false, neither the player or other code can get the character to move. AllowPlayerControl determines if the player can move the character. External  code can still make the character move if it is false